+++ trunk/src/blockmode.c
@@ -42,7 +42,7 @@ enum Blockmode_mode {
 	BLOCKMODE_HIGH = +1
 };
 
-struct BIN_MODE_INFO {	/* Usd for histogram binning */
+struct BIN_MODE_INFO {	/* Used for histogram binning */
 	double width;		/* The binning width used */
 	double i_offset;	/* 0.5 if we are to bin using the center the bins on multiples of width, else 0.0 */
 	double o_offset;	/* 0.0 if we are to report center the bins on multiples of width, else 0.5 */
@@ -210,8 +210,8 @@ struct BIN_MODE_INFO *bin_setup (struct GMT_CTRL *GMT, struct BLK_DATA *d, doubl
 	B->o_offset = (center) ? 0.0 : 0.5;
 	B->width = width;
 	B->i_width = 1.0 / width;
-	B->min = lrint (floor ((d[0].a[k]   * B->i_width) + B->i_offset));
-	B->max = lrint (ceil  ((d[n-1].a[k] * B->i_width) + B->i_offset));
+	B->min = (int)lrint (floor ((d[0].a[k]   * B->i_width) + B->i_offset));
+	B->max = (int)lrint (ceil  ((d[n-1].a[k] * B->i_width) + B->i_offset));
 	B->n_bins = B->max - B->min + 1;
 	B->count = GMT_memory (GMT, NULL, B->n_bins, unsigned int);
 	B->mode_choice = mode_choice;
@@ -233,7 +233,7 @@ double bin_mode (struct GMT_CTRL *GMT, struct BLK_DATA *d, uint64_t n, uint64_t 
 
 	GMT_memset (B->count, B->n_bins, unsigned int);	/* Reset the counts */
 	for (i = 0; i < n; i++) {	/* Loop over sorted data points */
-		bin = lrint (floor ((d[i].a[k] * B->i_width) + B->i_offset)) - B->min;
+		bin = (unsigned int)lrint (floor ((d[i].a[k] * B->i_width) + B->i_offset)) - B->min;
 		B->count[bin]++;
 		if (B->count[bin] > mode_count) {	/* New max count value; make a note */
 			mode_count = B->count[bin];	/* Highest count so far... */
